requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0
orjson>=3.9
rich>=13.7

# NLP / Retrieval
transformers>=4.41
torch>=2.2
sentence-transformers>=3.0
faiss-cpu>=1.8.0

# Datasets
datasets>=2.19

# Matbench (materials)
# matbench>=0.6.0
# pymatgen>=2024.5.1

